Tests/LibWeb/Text/expected/get-bounding-client-rect-pinch-zoom.txt

before: 0.000,0.000 100.000x100.000
visualViewport scale: 1.500
after: 0.000,0.000 100.000x100.000
size unchanged: PASS
